// frontend/contracts/core/v1/models/PostModel.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type PostModel = {
    id: string;
    created_at: string;
    updated_at: string;
    title: string;
    text: string;
    description: string;
    preview_image_id: string;
    author_id: string;
    author_username: string;
};